+++ src/main/java/uk/gov/companieshouse/api/accounts/model/entity/smallfull/notes/creditorswithinoneyear/CreditorsWithinOneYearDataEntity.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.api.accounts.model.entity.notes.creditorswithinoneyear;
+package uk.gov.companieshouse.api.accounts.model.entity.smallfull.notes.creditorswithinoneyear;
 
 import com.google.gson.Gson;
 import org.springframework.data.mongodb.core.mapping.Field;
