+++ src/main/java/uk/gov/companieshouse/api/accounts/model/entity/smallfull/notes/creditorswithinoneyear/CreditorsWithinOneYearEntity.java
@@ -1,13 +1,11 @@
-package uk.gov.companieshouse.api.accounts.model.entity.notes.creditorswithinoneyear;
+package uk.gov.companieshouse.api.accounts.model.entity.smallfull.notes.creditorswithinoneyear;
 
 import com.google.gson.Gson;
-import org.springframework.data.mongodb.core.mapping.Document;
-import uk.gov.companieshouse.api.accounts.model.entity.BaseEntity;
+import uk.gov.companieshouse.api.accounts.model.entity.NoteEntity;
 
 import java.util.Objects;
 
-@Document(collection = "notes")
-public class CreditorsWithinOneYearEntity extends BaseEntity {
+public class CreditorsWithinOneYearEntity extends NoteEntity {
 
     private CreditorsWithinOneYearDataEntity data;
