+++ src/main/java/uk/gov/companieshouse/api/accounts/model/rest/smallfull/notes/creditorswithinoneyear/PreviousPeriod.java
@@ -1,35 +1,44 @@
-package uk.gov.companieshouse.api.accounts.model.entity.notes.creditorswithinoneyear;
+package uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.creditorswithinoneyear;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.gson.Gson;
-import org.springframework.data.mongodb.core.mapping.Field;
+import org.hibernate.validator.constraints.Range;
 
 import java.util.Objects;
 
-public class CurrentPeriodEntity {
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class PreviousPeriod {
 
-    @Field("accruals_and_deferred_income")
+    private static final int MAX_RANGE = 99999999;
+    private static final int MIN_RANGE = 0;
+
+    @Range(min=MIN_RANGE,max=MAX_RANGE, message = "value.outside.range")
+    @JsonProperty("accruals_and_deferred_income")
     private Long accrualsAndDeferredIncome;
 
-    @Field("bank_loans_and_overdrafts")
+    @Range(min=MIN_RANGE,max=MAX_RANGE, message = "value.outside.range")
+    @JsonProperty("bank_loans_and_overdrafts")
     private Long bankLoansAndOverdrafts;
 
-    @Field("finance_leases_and_hire_purchase_contracts")
+    @Range(min=MIN_RANGE,max=MAX_RANGE, message = "value.outside.range")
+    @JsonProperty("finance_leases_and_hire_purchase_contracts")
     private Long financeLeasesAndHirePurchaseContracts;
 
-    @Field("other_creditors")
+    @Range(min=MIN_RANGE,max=MAX_RANGE, message = "value.outside.range")
+    @JsonProperty("other_creditors")
     private Long otherCreditors;
 
-    @Field("taxation_and_social_security")
+    @Range(min=MIN_RANGE,max=MAX_RANGE, message = "value.outside.range")
+    @JsonProperty("taxation_and_social_security")
     private Long taxationAndSocialSecurity;
 
-    @Field("trade_creditors")
+    @Range(min=MIN_RANGE,max=MAX_RANGE, message = "value.outside.range")
+    @JsonProperty("trade_creditors")
     private Long tradeCreditors;
 
-    @Field("total")
+    @JsonProperty("total")
     private Long total;
-
-    @Field("details")
-    private String details;
 
     public Long getAccrualsAndDeferredIncome() {
         return accrualsAndDeferredIncome;
@@ -87,27 +96,18 @@ public class CurrentPeriodEntity {
         this.total = total;
     }
 
-    public String getDetails() {
-        return details;
-    }
-
-    public void setDetails(String details) {
-        this.details = details;
-    }
-
     @Override
     public boolean equals(Object o) {
         if (this == o) {return true;}
-        if (!(o instanceof CurrentPeriodEntity)) {return false;}
-        CurrentPeriodEntity that = (CurrentPeriodEntity) o;
+        if (!(o instanceof PreviousPeriod)) {return false;}
+        PreviousPeriod that = (PreviousPeriod) o;
         return Objects.equals(getAccrualsAndDeferredIncome(), that.getAccrualsAndDeferredIncome()) &&
             Objects.equals(getBankLoansAndOverdrafts(), that.getBankLoansAndOverdrafts()) &&
             Objects.equals(getFinanceLeasesAndHirePurchaseContracts(), that.getFinanceLeasesAndHirePurchaseContracts()) &&
             Objects.equals(getOtherCreditors(), that.getOtherCreditors()) &&
             Objects.equals(getTaxationAndSocialSecurity(), that.getTaxationAndSocialSecurity()) &&
             Objects.equals(getTradeCreditors(), that.getTradeCreditors()) &&
-            Objects.equals(getTotal(), that.getTotal()) &&
-            Objects.equals(getDetails(), that.getDetails());
+            Objects.equals(getTotal(), that.getTotal());
     }
 
     @Override
@@ -115,7 +115,7 @@ public class CurrentPeriodEntity {
 
         return Objects.hash(getAccrualsAndDeferredIncome(), getBankLoansAndOverdrafts(),
             getFinanceLeasesAndHirePurchaseContracts(), getOtherCreditors(), getTaxationAndSocialSecurity(),
-            getTradeCreditors(), getTotal(), getDetails());
+            getTradeCreditors(), getTotal());
     }
 
     @Override
