+++ src/main/java/uk/gov/companieshouse/api/accounts/repository/smallfull/CreditorsWithinOneYearRepository.java
@@ -1,0 +1,16 @@
+package uk.gov.companieshouse.api.accounts.repository.smallfull;
+
+import org.springframework.stereotype.Repository;
+import uk.gov.companieshouse.api.accounts.enumeration.AccountingNoteType;
+import uk.gov.companieshouse.api.accounts.model.entity.smallfull.notes.creditorswithinoneyear.CreditorsWithinOneYearEntity;
+import uk.gov.companieshouse.api.accounts.repository.AccountsNoteRepository;
+
+@Repository
+public interface CreditorsWithinOneYearRepository extends AccountsNoteRepository<CreditorsWithinOneYearEntity> {
+
+    @Override
+    default AccountingNoteType getAccountsNote() {
+        return AccountingNoteType.SMALL_FULL_CREDITORS_WITHIN;
+    }
+
+}
