+++ src/main/java/uk/gov/companieshouse/api/accounts/validation/CreditorsWithinOneYearValidator.java
@@ -2,13 +2,14 @@ package uk.gov.companieshouse.api.accounts.validation;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import uk.gov.companieshouse.api.accounts.enumeration.AccountingNoteType;
 import uk.gov.companieshouse.api.accounts.exception.DataException;
 import uk.gov.companieshouse.api.accounts.exception.ServiceException;
 import uk.gov.companieshouse.api.accounts.model.rest.BalanceSheet;
 import uk.gov.companieshouse.api.accounts.model.rest.OtherLiabilitiesOrAssets;
-import uk.gov.companieshouse.api.accounts.model.rest.notes.creditorswithinoneyear.CreditorsWithinOneYear;
-import uk.gov.companieshouse.api.accounts.model.rest.notes.creditorswithinoneyear.CurrentPeriod;
-import uk.gov.companieshouse.api.accounts.model.rest.notes.creditorswithinoneyear.PreviousPeriod;
+import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.creditorswithinoneyear.CreditorsWithinOneYear;
+import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.creditorswithinoneyear.CurrentPeriod;
+import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.creditorswithinoneyear.PreviousPeriod;
 import uk.gov.companieshouse.api.accounts.model.validation.Errors;
 import uk.gov.companieshouse.api.accounts.service.CompanyService;
 import uk.gov.companieshouse.api.accounts.service.impl.CurrentPeriodService;
@@ -21,7 +22,7 @@ import javax.validation.Valid;
 import java.util.Optional;
 
 @Component
-public class CreditorsWithinOneYearValidator extends BaseValidator implements CrossValidator<CreditorsWithinOneYear> {
+public class CreditorsWithinOneYearValidator extends BaseValidator implements NoteValidator<CreditorsWithinOneYear> {
 
     private static final String CREDITORS_WITHIN_PATH = "$.creditors_within_one_year";
     private static final String CREDITORS_WITHIN_CURRENT_PERIOD_PATH = CREDITORS_WITHIN_PATH +
@@ -46,20 +47,6 @@ public class CreditorsWithinOneYearValidator extends BaseValidator implements Cr
         this.previousPeriodService = previousPeriodService;
     }
 
-    public boolean validateIfOnlyDetails(CurrentPeriod currentPeriodNote) {
-        return (currentPeriodNote != null && currentPeriodNote.getDetails() != null
-                && creditorsNumberFieldsAreNull(currentPeriodNote));
-    }
-
-    private boolean creditorsNumberFieldsAreNull(CurrentPeriod currentPeriodNote) {
-        return currentPeriodNote.getTotal() == null && currentPeriodNote.getTradeCreditors() == null &&
-                currentPeriodNote.getAccrualsAndDeferredIncome() == null &&
-                currentPeriodNote.getOtherCreditors() == null &&
-                currentPeriodNote.getTaxationAndSocialSecurity() == null &&
-                currentPeriodNote.getFinanceLeasesAndHirePurchaseContracts() == null &&
-                currentPeriodNote.getBankLoansAndOverdrafts() == null;
-    }
-
     private Errors validateIfEmptyResource(CreditorsWithinOneYear creditorsWithinOneYear,
             HttpServletRequest request, String companyAccountsId) throws DataException {
 
@@ -80,11 +67,8 @@ public class CreditorsWithinOneYearValidator extends BaseValidator implements Cr
         return errors;
     }
 
-    public Errors validateCreditorsWithinOneYear(@Valid CreditorsWithinOneYear creditorsWithinOneYear,
-            Transaction transaction,
-            String companyAccountsId,
-            HttpServletRequest request) throws DataException {
-
+    @Override
+    public Errors validateSubmission(CreditorsWithinOneYear creditorsWithinOneYear, Transaction transaction, String companyAccountsId, HttpServletRequest request) throws DataException {
         Errors errors = validateIfEmptyResource(creditorsWithinOneYear, request, companyAccountsId);
 
         if (errors.hasErrors()) {
@@ -279,25 +263,6 @@ public class CreditorsWithinOneYearValidator extends BaseValidator implements Cr
                 , errors);
     }
 
-    @Override
-    public Errors crossValidate(CreditorsWithinOneYear creditorsWithinOneYear,
-            HttpServletRequest request,
-            String companyAccountsId,
-            Errors errors) throws DataException {
-
-        BalanceSheet currentPeriodBalanceSheet = getCurrentPeriodBalanceSheet(request,
-                companyAccountsId);
-        BalanceSheet previousPeriodBalanceSheet = getPreviousPeriodBalanceSheet(request,
-                companyAccountsId);
-
-        crossValidateCurrentPeriodFields(creditorsWithinOneYear.getCurrentPeriod(),
-                currentPeriodBalanceSheet, errors);
-        crossValidatePreviousPeriodFields(creditorsWithinOneYear.getPreviousPeriod(),
-                previousPeriodBalanceSheet, errors);
-
-        return errors;
-    }
-
     private void crossValidatePreviousPeriodFields(PreviousPeriod previousPeriodNote,
             BalanceSheet previousPeriodBalanceSheet,
             Errors errors) {
@@ -457,5 +422,10 @@ public class CreditorsWithinOneYearValidator extends BaseValidator implements Cr
         return ! Optional.ofNullable(previousPeriodNote)
                 .map(PreviousPeriod :: getTotal)
                 .isPresent();
+    }
+
+    @Override
+    public AccountingNoteType getAccountingNoteType() {
+        return AccountingNoteType.SMALL_FULL_CREDITORS_WITHIN;
     }
 }
