+++ src/main/resources/notes.yml
@@ -7,3 +7,4 @@ notes:
     - debtors
     - current-assets-investments
     - intangible-assets
+    - creditors-within-one-year
