+++ src/test/java/uk/gov/companieshouse/api/accounts/validation/CreditorsWithinOneYearValidatorTests.java
@@ -16,13 +16,14 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.test.util.ReflectionTestUtils;
+import uk.gov.companieshouse.api.accounts.enumeration.AccountingNoteType;
 import uk.gov.companieshouse.api.accounts.exception.DataException;
 import uk.gov.companieshouse.api.accounts.exception.ServiceException;
 import uk.gov.companieshouse.api.accounts.model.rest.BalanceSheet;
 import uk.gov.companieshouse.api.accounts.model.rest.OtherLiabilitiesOrAssets;
-import uk.gov.companieshouse.api.accounts.model.rest.notes.creditorswithinoneyear.CreditorsWithinOneYear;
-import uk.gov.companieshouse.api.accounts.model.rest.notes.creditorswithinoneyear.CurrentPeriod;
-import uk.gov.companieshouse.api.accounts.model.rest.notes.creditorswithinoneyear.PreviousPeriod;
+import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.creditorswithinoneyear.CreditorsWithinOneYear;
+import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.creditorswithinoneyear.CurrentPeriod;
+import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.creditorswithinoneyear.PreviousPeriod;
 import uk.gov.companieshouse.api.accounts.model.validation.Error;
 import uk.gov.companieshouse.api.accounts.model.validation.Errors;
 import uk.gov.companieshouse.api.accounts.service.CompanyService;
@@ -101,7 +102,7 @@ public class CreditorsWithinOneYearValidatorTests {
 
         mockValidBalanceSheetCurrentPeriod();
 
-        errors = validator.validateCreditorsWithinOneYear(creditorsWithinOneYear, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest);
+        errors = validator.validateSubmission(creditorsWithinOneYear, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest);
 
         assertFalse(errors.hasErrors());
     }
@@ -118,7 +119,7 @@ public class CreditorsWithinOneYearValidatorTests {
 
         when(mockCompanyService.isMultipleYearFiler(mockTransaction)).thenReturn(true);
 
-        errors = validator.validateCreditorsWithinOneYear(creditorsWithinOneYear, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest);
+        errors = validator.validateSubmission(creditorsWithinOneYear, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest);
 
         assertFalse(errors.hasErrors());
     }
@@ -129,7 +130,7 @@ public class CreditorsWithinOneYearValidatorTests {
 
         createValidNoteCurrentPeriod();
 
-        errors = validator.validateCreditorsWithinOneYear(creditorsWithinOneYear, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest);
+        errors = validator.validateSubmission(creditorsWithinOneYear, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest);
 
         assertFalse(errors.hasErrors());
     }
@@ -143,35 +144,7 @@ public class CreditorsWithinOneYearValidatorTests {
 
         when(mockCompanyService.isMultipleYearFiler(mockTransaction)).thenReturn(true);
 
-        errors = validator.validateCreditorsWithinOneYear(creditorsWithinOneYear, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest);
-
-        assertFalse(errors.hasErrors());
-    }
-
-    @Test
-    @DisplayName("Cross validation passes with valid note for first year filer")
-    void testSuccessfulCrossValidationForFirstYearFiler() throws DataException {
-
-        createValidNoteCurrentPeriod();
-
-        mockValidBalanceSheetCurrentPeriod();
-
-        errors = validator.crossValidate(creditorsWithinOneYear, mockRequest, COMPANY_ACCOUNTS_ID, errors);
-
-        assertFalse(errors.hasErrors());
-    }
-
-    @Test
-    @DisplayName("Cross validation passes with valid note for multiple year filer")
-    void testSuccessfulCrossValidationForMultipleYearFiler() throws DataException {
-
-        createValidNoteCurrentPeriod();
-        createValidNotePreviousPeriod();
-
-        mockValidBalanceSheetCurrentPeriod();
-        mockValidBalanceSheetPreviousPeriod();
-
-        errors = validator.crossValidate(creditorsWithinOneYear, mockRequest, COMPANY_ACCOUNTS_ID, errors);
+        errors = validator.validateSubmission(creditorsWithinOneYear, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest);
 
         assertFalse(errors.hasErrors());
     }
@@ -188,7 +161,7 @@ public class CreditorsWithinOneYearValidatorTests {
 
         when(mockCompanyService.isMultipleYearFiler(mockTransaction)).thenReturn(true);
 
-        errors = validator.validateCreditorsWithinOneYear(creditorsWithinOneYear, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest);
+        errors = validator.validateSubmission(creditorsWithinOneYear, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest);
 
         assertEquals(2, errors.getErrorCount());
         assertTrue(errors.containsError(createError(MANDATORY_ELEMENT_MISSING_VALUE,
@@ -211,7 +184,7 @@ public class CreditorsWithinOneYearValidatorTests {
 
         when(mockCompanyService.isMultipleYearFiler(mockTransaction)).thenReturn(false);
 
-        errors = validator.validateCreditorsWithinOneYear(creditorsWithinOneYear, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest);
+        errors = validator.validateSubmission(creditorsWithinOneYear, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest);
 
         assertEquals(1, errors.getErrorCount());
         assertTrue(errors.containsError(createError(UNEXPECTED_DATA_VALUE,
@@ -243,7 +216,7 @@ public class CreditorsWithinOneYearValidatorTests {
         ReflectionTestUtils.setField(validator, MANDATORY_ELEMENT_MISSING_NAME,
                 MANDATORY_ELEMENT_MISSING_VALUE);
 
-        errors = validator.validateCreditorsWithinOneYear(creditorsWithinOneYear, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest);
+        errors = validator.validateSubmission(creditorsWithinOneYear, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest);
 
         assertTrue(errors.hasErrors());
         assertEquals(4, errors.getErrorCount());
@@ -283,7 +256,7 @@ public class CreditorsWithinOneYearValidatorTests {
         ReflectionTestUtils.setField(validator, PREVIOUS_BALANCE_SHEET_NOT_EQUAL_NAME,
                 PREVIOUS_BALANCE_SHEET_NOT_EQUAL_VALUE);
 
-        errors = validator.validateCreditorsWithinOneYear(creditorsWithinOneYear, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest);
+        errors = validator.validateSubmission(creditorsWithinOneYear, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest);
 
         assertTrue(errors.hasErrors());
         assertEquals(4, errors.getErrorCount());
@@ -318,7 +291,7 @@ public class CreditorsWithinOneYearValidatorTests {
         ReflectionTestUtils.setField(validator, UNEXPECTED_DATA_NAME,
                 UNEXPECTED_DATA_VALUE);
 
-        errors = validator.validateCreditorsWithinOneYear(creditorsWithinOneYear, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest);
+        errors = validator.validateSubmission(creditorsWithinOneYear, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest);
 
         assertTrue(errors.hasErrors());
         assertEquals(2, errors.getErrorCount());
@@ -347,7 +320,7 @@ public class CreditorsWithinOneYearValidatorTests {
 
         when(mockCompanyService.isMultipleYearFiler(mockTransaction)).thenReturn(true);
 
-        errors = validator.validateCreditorsWithinOneYear(creditorsWithinOneYear, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest);
+        errors = validator.validateSubmission(creditorsWithinOneYear, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest);
 
         assertTrue(errors.hasErrors());
         assertEquals(4, errors.getErrorCount());
@@ -373,7 +346,7 @@ public class CreditorsWithinOneYearValidatorTests {
         when(mockCompanyService.isMultipleYearFiler(mockTransaction)).thenThrow(mockServiceException);
 
         assertThrows(DataException.class,
-                () -> validator.validateCreditorsWithinOneYear(creditorsWithinOneYear,
+                () -> validator.validateSubmission(creditorsWithinOneYear,
                         mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest));
     }
 
@@ -388,7 +361,7 @@ public class CreditorsWithinOneYearValidatorTests {
         when(mockCurrentPeriodService.find(COMPANY_ACCOUNTS_ID, mockRequest)).thenThrow(new DataException(""));
 
         assertThrows(DataException.class,
-                () -> validator.validateCreditorsWithinOneYear(creditorsWithinOneYear,
+                () -> validator.validateSubmission(creditorsWithinOneYear,
                         mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest));
     }
 
@@ -405,8 +378,15 @@ public class CreditorsWithinOneYearValidatorTests {
         when(mockPreviousPeriodService.find(COMPANY_ACCOUNTS_ID, mockRequest)).thenThrow(new DataException(""));
 
         assertThrows(DataException.class,
-                () -> validator.validateCreditorsWithinOneYear(creditorsWithinOneYear,
+                () -> validator.validateSubmission(creditorsWithinOneYear,
                         mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest));
+    }
+
+    @Test
+    @DisplayName("Get accounting note type")
+    void getAccountingNoteType() {
+
+        assertEquals(AccountingNoteType.SMALL_FULL_CREDITORS_WITHIN, validator.getAccountingNoteType());
     }
 
     private void createValidNoteCurrentPeriod() {
